Automate process to build against both 10.8.5 and 10.1.1 SDKs
Port the code over from ltn-lted-build so that we can build
multiple versions of the OBE binary based on the decklink SDK
version.  Note that this includes a bit of platform-specific code
and OBE doesn't actually build on OSX, but this is just a couple
of extra lines and makes the logic consistent with the VLC and LTED
build processes.